Fixed Reset Bug (#1)
In SharedCommunications.c the function setRemotePower used the function
utoa(). The buffer receiving the output was 5 - too short for storing
any potential integer value. This has been changed to 7 to store
"-32768" plus null termination

In SharedCommunications.c the commandBuilder functions use the macro
RECEIVE_PARAM_LENGTH to set a buffer length of 20.  This appeared to
cause a crash, and likely so since some of the commands could extend
beyond 20 characters. This macro has been changed to 50 in two places