Reset Time
Cleaned up reset time menu display to be a bit more readable
It is OK for the reset time to be not on a 30 minute mark.
This menu will display it as it is.
But if it is changed from this menu it is restricted to the 30 marks
Had to change the math in the menu to support to non 30 minute start times